creating generic reporter for test command, and refactoring legacy reporting behaviour to this struct
