+++ guard/src/commands/test.rs
@@ -1,14 +1,15 @@
-use crate::commands::reporters::test::{ContextAwareRule, Err, StructuredTestReporter, TestResult};
+use crate::commands::reporters::test::generic::GenericReporter;
+use crate::commands::reporters::test::structured::{
+    ContextAwareRule, Err, StructuredTestReporter, TestResult,
+};
 use crate::commands::reporters::JunitReport;
 use crate::commands::{SUCCESS_STATUS_CODE, TEST_ERROR_STATUS_CODE, TEST_FAILURE_STATUS_CODE};
 use clap::{Arg, ArgAction, ArgGroup, ArgMatches, ValueHint};
 use serde::{Deserialize, Serialize};
 use std::collections::{BTreeMap, HashMap};
-use std::convert::TryFrom;
 use std::fs::File;
 use std::io::Write;
 use std::path::{Path, PathBuf};
-use std::rc::Rc;
 use std::time::Instant;
 use walkdir::DirEntry;
 
@@ -16,8 +17,7 @@ use validate::validate_path;
 
 use crate::command::Command;
 use crate::commands::files::{
-    alpabetical, get_files_with_filter, iterate_over, last_modified, read_file_content,
-    regular_ordering,
+    alpabetical, get_files_with_filter, last_modified, read_file_content, regular_ordering,
 };
 use crate::commands::validate::{OutputFormatType, OUTPUT_FORMAT_VALUE_TYPE};
 use crate::commands::{
@@ -25,11 +25,7 @@ use crate::commands::{
     RULES_AND_TEST_FILE, RULES_FILE, TEST, TEST_DATA, VERBOSE,
 };
 use crate::rules::errors::Error;
-use crate::rules::eval::eval_rules_file;
-use crate::rules::exprs::RulesFile;
-use crate::rules::path_value::PathAwareValue;
-use crate::rules::Status::SKIP;
-use crate::rules::{NamedStatus, RecordType, Result, Status};
+use crate::rules::Result;
 use crate::utils::reader::Reader;
 use crate::utils::writer::Writer;
 
@@ -274,7 +270,15 @@ fn handle_plaintext_directory(
                         .iter()
                         .map(|de| de.path().to_path_buf())
                         .collect::<Vec<PathBuf>>();
-                    let test_exit_code = test_with_data(&data_test_files, &rules, verbose, writer)?;
+
+                    let mut reporter = GenericReporter {
+                        test_data: &data_test_files,
+                        rules,
+                        verbose,
+                        writer,
+                    };
+
+                    let test_exit_code = reporter.report()?;
 
                     exit_code = if exit_code == 0 {
                         test_exit_code
@@ -310,7 +314,17 @@ fn handle_plaintext_single_file(
                     writeln!(writer, "Parse Error on ruleset file {e}")?;
                     Ok(TEST_ERROR_STATUS_CODE)
                 }
-                Ok(Some(rules)) => test_with_data(data_test_files, &rules, verbose, writer),
+
+                Ok(Some(rules)) => {
+                    let mut reporter = GenericReporter {
+                        test_data: data_test_files,
+                        writer,
+                        verbose,
+                        rules,
+                    };
+
+                    reporter.report()
+                }
                 Ok(None) => Ok(SUCCESS_STATUS_CODE),
             }
         }
@@ -370,23 +384,7 @@ pub(crate) fn handle_structured_single_report(
     match output {
         OutputFormatType::YAML => serde_yaml::to_writer(writer, &result)?,
         OutputFormatType::JSON => serde_json::to_writer_pretty(writer, &result)?,
-        OutputFormatType::Junit => {
-            let suite = result.build_test_suite();
-            let failures = suite.failures;
-            let tests = suite.test_cases.len();
-            let errors = suite.errors;
-            let time = suite.time;
-
-            JunitReport {
-                name: "cfn-guard test report",
-                test_suites: vec![suite],
-                failures,
-                errors,
-                tests,
-                duration: time,
-            }
-            .serialize(writer)?;
-        }
+        OutputFormatType::Junit => JunitReport::from(&vec![result]).serialize(writer)?,
         OutputFormatType::SingleLineSummary => unreachable!(),
     }
 
@@ -460,35 +458,7 @@ fn handle_structured_directory_report(
     match output {
         OutputFormatType::YAML => serde_yaml::to_writer(writer, &test_results)?,
         OutputFormatType::JSON => serde_json::to_writer_pretty(writer, &test_results)?,
-        OutputFormatType::Junit => {
-            let mut errors = 0;
-            let mut failures = 0;
-            let mut tests = 0;
-            let mut time = 0;
-
-            let test_suites = test_results.iter().fold(vec![], |mut acc, result| {
-                let suite = result.build_test_suite();
-
-                time += suite.time;
-                errors += suite.errors;
-                failures += suite.failures;
-                tests += suite.test_cases.len();
-
-                acc.push(suite);
-                acc
-            });
-
-            let report = JunitReport {
-                name: "cfn-guard test report",
-                test_suites,
-                failures,
-                errors,
-                tests,
-                duration: time,
-            };
-
-            report.serialize(writer)?;
-        }
+        OutputFormatType::Junit => JunitReport::from(&test_results).serialize(writer)?,
         // NOTE: safe since output type is checked prior to calling this function
         OutputFormatType::SingleLineSummary => unreachable!(),
     }
@@ -521,139 +491,6 @@ pub struct TestSpec {
     pub name: Option<String>,
     pub input: serde_yaml::Value,
     pub expectations: TestExpectations,
-}
-
-#[allow(clippy::never_loop)]
-fn test_with_data(
-    test_data_files: &[PathBuf],
-    rules: &RulesFile<'_>,
-    verbose: bool,
-    writer: &mut Writer,
-) -> Result<i32> {
-    let mut exit_code = 0;
-    let mut test_counter = 1;
-    for specs in iterate_over(test_data_files, |data, path| {
-        match serde_yaml::from_str::<Vec<TestSpec>>(&data) {
-            Ok(spec) => Ok(spec),
-            Err(_) => match serde_json::from_str::<Vec<TestSpec>>(&data) {
-                Ok(specs) => Ok(specs),
-                Err(e) => Err(Error::ParseError(format!(
-                    "Unable to process data in file {}, Error {},",
-                    path.display(),
-                    e
-                ))),
-            },
-        }
-    }) {
-        match specs {
-            Err(e) => {
-                writeln!(writer, "Error processing {e}")?;
-                exit_code = TEST_ERROR_STATUS_CODE;
-            }
-            Ok(specs) => {
-                for each in specs {
-                    writeln!(writer, "Test Case #{test_counter}")?;
-                    if each.name.is_some() {
-                        writeln!(writer, "Name: {}", each.name.unwrap_or_default())?;
-                    }
-
-                    let by_result = {
-                        let mut by_result = HashMap::new();
-                        let root = PathAwareValue::try_from(each.input)?;
-
-                        let mut root_scope =
-                            crate::rules::eval_context::root_scope(rules, Rc::new(root));
-
-                        eval_rules_file(rules, &mut root_scope, None)?; // we never use data file name in the output
-
-                        let top = root_scope.reset_recorder().extract();
-
-                        let by_rules: HashMap<&str, Vec<&Option<RecordType<'_>>>> =
-                            top.children.iter().fold(HashMap::new(), |mut acc, rule| {
-                                if let Some(RecordType::RuleCheck(NamedStatus { name, .. })) =
-                                    rule.container
-                                {
-                                    acc.entry(name).or_default().push(&rule.container)
-                                }
-                                acc
-                            });
-
-                        for (rule_name, rule) in by_rules {
-                            let expected = match each.expectations.rules.get(rule_name) {
-                                Some(exp) => Status::try_from(exp.as_str())?,
-                                None => {
-                                    writeln!(
-                                        writer,
-                                        "  No Test expectation was set for Rule {rule_name}"
-                                    )?;
-                                    continue;
-                                }
-                            };
-
-                            let mut statues: Vec<Status> = Vec::with_capacity(rule.len());
-                            let matched = 'matched: loop {
-                                let mut all_skipped = 0;
-
-                                for each in rule.iter().copied().flatten() {
-                                    if let RecordType::RuleCheck(NamedStatus {
-                                        status: got_status,
-                                        ..
-                                    }) = each
-                                    {
-                                        match expected {
-                                            SKIP => {
-                                                if *got_status == SKIP {
-                                                    all_skipped += 1;
-                                                }
-                                            }
-
-                                            rest => {
-                                                if *got_status == rest {
-                                                    break 'matched Some(expected);
-                                                }
-                                            }
-                                        }
-                                        statues.push(*got_status)
-                                    }
-                                }
-                                if expected == SKIP && all_skipped == rule.len() {
-                                    break 'matched Some(expected);
-                                }
-                                break 'matched None;
-                            };
-
-                            match matched {
-                                Some(status) => {
-                                    by_result
-                                        .entry(String::from("PASS"))
-                                        .or_insert_with(indexmap::IndexSet::new)
-                                        .insert(format!("{rule_name}: Expected = {status}"));
-                                }
-
-                                None => {
-                                    by_result
-                                        .entry(String::from("FAIL"))
-                                        .or_insert_with(indexmap::IndexSet::new)
-                                        .insert(format!(
-                                            "{rule_name}: Expected = {expected}, Evaluated = {statues:?}"
-                                        ));
-                                    exit_code = TEST_FAILURE_STATUS_CODE;
-                                }
-                            }
-                        }
-
-                        if verbose {
-                            validate::print_verbose_tree(&top, writer);
-                        }
-                        by_result
-                    };
-                    print_test_case_report(&by_result, writer);
-                    test_counter += 1;
-                }
-            }
-        }
-    }
-    Ok(exit_code)
 }
 
 struct OrderedTestDirectory(BTreeMap<String, Vec<GuardFile>>);
@@ -740,22 +577,4 @@ impl From<walkdir::WalkDir> for OrderedTestDirectory {
 
         OrderedTestDirectory(files)
     }
-}
-
-pub(crate) fn print_test_case_report(
-    by_result: &HashMap<String, indexmap::IndexSet<String>>,
-    writer: &mut Writer,
-) {
-    use itertools::Itertools;
-    let mut results = by_result.keys().cloned().collect_vec();
-
-    results.sort(); // Deterministic order of results
-
-    for result in &results {
-        writeln!(writer, "  {result} Rules:").expect("Unable to write to the output");
-        for each_case in by_result.get(result).unwrap() {
-            writeln!(writer, "    {}", *each_case).expect("Unable to write to the output");
-        }
-    }
-    writeln!(writer).expect("Unable to write to the output");
 }
